Popover Component
The popover component is a utility component that requires an anchor
in the form of a DOM node or a jquery selector. Optionally, it takes
a preferred position for rendering relative to the anchor.

It encapsulates the headaches and edge cases of positioning the popover
pane always in the window with the arrow always pointing at the anchor.